api/requirements.txt
# FastAPI and web framework
fastapi>=0.104.0,<0.105.0
uvicorn[standard]>=0.24.0,<0.25.0
python-multipart>=0.0.6,<0.1.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# Database
sqlalchemy>=2.0.0,<3.0.0
alembic>=1.12.0,<2.0.0
asyncpg>=0.29.0,<1.0.0
psycopg2-binary>=2.9.0,<3.0.0
pgvector>=0.2.0,<1.0.0

# Authentication
supabase>=2.0.0,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[argon2]>=1.7.4,<2.0.0
argon2-cffi>=23.1.0,<24.0.0

# Storage
boto3>=1.29.0,<2.0.0
minio>=7.2.0,<8.0.0

# Redis and task queue
redis>=5.0.0,<6.0.0
hiredis>=2.2.0,<3.0.0
dramatiq[redis]>=1.15.0,<2.0.0

# Validation and utilities
email-validator>=2.1.0,<3.0.0
python-magic>=0.4.27,<1.0.0
filetype>=1.2.0,<2.0.0

# Rate limiting
slowapi>=0.1.9,<1.0.0

# Observability
opentelemetry-api>=1.21.0,<2.0.0
opentelemetry-sdk>=1.21.0,<2.0.0
opentelemetry-instrumentation-fastapi>=0.42b0,<1.0.0
opentelemetry-exporter-otlp>=1.21.0,<2.0.0
prometheus-client>=0.19.0,<1.0.0
structlog>=23.2.0,<24.0.0

# HTTP client
httpx>=0.25.0,<1.0.0

# Development
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0
pytest-cov>=4.1.0,<5.0.0
black>=23.11.0,<24.0.0
ruff>=0.1.6,<1.0.0
mypy>=1.7.0,<2.0.0

# Security
bandit>=1.7.5,<2.0.0
safety>=2.3.0,<3.0.0
